R CMD check

+++ src/sentencepiece/src/char_model_trainer.cc
@@ -37,7 +37,7 @@ util::Status Trainer::Train() {
     sum += it.second;
   }
 
-  const auto logsum = std::log(static_cast<float>(sum));
+  const auto logsum = static_cast<float>(log(static_cast<long double>(sum)));
 
   CHECK_OR_RETURN(final_pieces_.empty());
   for (const auto &it : Sorted(required_chars_)) {
@@ -47,7 +47,7 @@ util::Status Trainer::Train() {
     }
     final_pieces_.emplace_back(
         string_util::UnicodeCharToUTF8(it.first),
-        std::log(static_cast<float>(it.second)) - logsum);
+        static_cast<float>(log(static_cast<long double>(it.second))) - logsum);
   }
 
   if (trainer_spec_.use_all_vocab()) {
